+++ changelog.md
@@ -1,2 +1,5 @@
 #### v0.1.1
-* First version of spoty-list-mix
+* First version of spoty-list-mix, only with search tracks and create playlist
+
+#### v0.1.0
+* First version in develop version of spoty-list-mix